requirements.txt
openai
pyaudio
requests
wave
elevenlabs
pydub
webrtcvad-wheels
python-dotenv
ffmpeg
flask